text-detection-python-tesseract-easyocr-textract/requirements.txt
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pytesseract==0.3.13
